refactor(native): modularize browser controls
