+++ src-tauri/src/browser.rs
@@ -38,81 +38,15 @@ const OFFSCREEN_Y: f64 = -10000.0;
 const USER_NAVIGATION_MARKER_TTL: Duration = Duration::from_secs(2);
 const MAX_TRACKED_BROWSER_URLS: usize = 64;
 
-#[derive(Debug, Clone, Copy, PartialEq)]
-enum BrowserBounds {
-    Hidden { w: f64, h: f64 },
-    Visible { x: f64, y: f64, w: f64, h: f64 },
-}
+mod browser_bounds;
+mod browser_commands;
+mod browser_native;
 
-fn browser_bounds_within_client(
-    client_w: f64,
-    client_h: f64,
-    x: f64,
-    y: f64,
-    w: f64,
-    h: f64,
-) -> Result<BrowserBounds, String> {
-    if !client_w.is_finite()
-        || !client_h.is_finite()
-        || !x.is_finite()
-        || !y.is_finite()
-        || !w.is_finite()
-        || !h.is_finite()
-    {
-        return Err("browser bounds must be finite".to_string());
-    }
-
-    let client_w = client_w.max(0.0);
-    let client_h = client_h.max(0.0);
-    let realized_w = w.max(2.0).min(client_w.max(1.0));
-    let realized_h = h.max(2.0).min(client_h.max(1.0));
-    if client_w <= 1.0 || client_h <= 1.0 || x < 0.0 || y < 0.0 || w <= 1.0 || h <= 1.0 {
-        return Ok(BrowserBounds::Hidden {
-            w: realized_w,
-            h: realized_h,
-        });
-    }
-
-    if x >= client_w - 1.0 || y >= client_h - 1.0 {
-        return Ok(BrowserBounds::Hidden {
-            w: realized_w,
-            h: realized_h,
-        });
-    }
-
-    let x = x.max(0.0);
-    let y = y.max(0.0);
-    let w = w.min(client_w - x);
-    let h = h.min(client_h - y);
-    if w <= 1.0 || h <= 1.0 {
-        return Ok(BrowserBounds::Hidden {
-            w: realized_w,
-            h: realized_h,
-        });
-    }
-    Ok(BrowserBounds::Visible { x, y, w, h })
-}
-
-/// Realize a new native browser child at its full intended size but outside
-/// the main window. WRY creates its `WRY_WEBVIEW` container HWND before
-/// WebView2 environment/controller initialization finishes. If that async COM
-/// callback stalls, Tauri has no registered Webview handle to hide yet; an
-/// in-viewport container would therefore remain above the app and intercept
-/// input. Creation is fail-closed offscreen and reconciliation seats the child
-/// only after `add_child` has returned and the newest intent is still visible.
-fn offscreen_browser_creation_bounds(
-    client_w: f64,
-    client_h: f64,
-    w: f64,
-    h: f64,
-) -> Result<(f64, f64), String> {
-    match browser_bounds_within_client(client_w, client_h, OFFSCREEN_X, OFFSCREEN_Y, w, h)? {
-        BrowserBounds::Hidden { w, h } => Ok((w, h)),
-        BrowserBounds::Visible { .. } => {
-            Err("offscreen browser creation unexpectedly produced visible bounds".to_string())
-        }
-    }
-}
+use browser_bounds::{
+    browser_bounds_within_client, offscreen_browser_creation_bounds, BrowserBounds,
+};
+pub use browser_commands::*;
+use browser_native::{hide_webview, show_webview_at};
 
 #[derive(Clone, Debug, PartialEq)]
 struct BrowserNavigationIntent {
@@ -990,58 +924,6 @@ impl EnvironmentCallbackTimeoutRetryState {
     }
 }
 
-// Park the webview offscreen at its CURRENT size. Do not shrink it to 1×1:
-// collapsing the layer lets WKWebView drop its backing surface, and a later
-// browser_set_bounds re-seat can land as an unpainted (black) layer. Keeping
-// the real size while offscreen keeps the layer realized so it repaints
-// immediately when shown again.
-fn hide_webview(webview: &tauri::Webview) -> Result<(), String> {
-    // Offscreen parking is not a visibility guarantee on Windows: WebView2
-    // can retain a stale native input surface and invisibly capture Cave
-    // clicks. Hide the child layer through the platform API instead.
-    #[cfg(target_os = "windows")]
-    webview.hide().map_err(|e| e.to_string())?;
-
-    // WKWebView may drop its backing surface when hidden, so other platforms
-    // retain the realized layer at its current size and move it offscreen.
-    #[cfg(not(target_os = "windows"))]
-    webview
-        .set_position(LogicalPosition::new(OFFSCREEN_X, OFFSCREEN_Y))
-        .map_err(|e| e.to_string())?;
-    Ok(())
-}
-
-fn show_webview_at(webview: &tauri::Webview, x: f64, y: f64, w: f64, h: f64) -> Result<(), String> {
-    // Clamp to the main client area and apply position+size atomically. Two
-    // dispatcher calls briefly expose an old-size/new-position WebView2 layer
-    // during resize, which can cover unrelated UI and capture its clicks.
-    let window = webview.window();
-    let scale = window.scale_factor().map_err(|e| e.to_string())?;
-    let client = window
-        .inner_size()
-        .map_err(|e| e.to_string())?
-        .to_logical::<f64>(scale);
-    let bounds = match browser_bounds_within_client(client.width, client.height, x, y, w, h) {
-        Ok(bounds) => bounds,
-        Err(error) => {
-            hide_webview(webview)?;
-            return Err(error);
-        }
-    };
-    let BrowserBounds::Visible { x, y, w, h } = bounds else {
-        return hide_webview(webview);
-    };
-    webview
-        .set_bounds(Rect {
-            position: LogicalPosition::new(x, y).into(),
-            size: LogicalSize::new(w, h).into(),
-        })
-        .map_err(|e| e.to_string())?;
-    #[cfg(target_os = "windows")]
-    webview.show().map_err(|e| e.to_string())?;
-    Ok(())
-}
-
 fn ensure_browser_controller(caller: &tauri::Webview) -> Result<(), String> {
     if caller.label() != "main" {
         return Err("native browser controls are restricted to the main webview".to_string());
@@ -1406,302 +1288,6 @@ fn schedule_scope_reconcile(
     for label in labels {
         schedule_browser_reconcile(app.clone(), state.clone(), label);
     }
-    Ok(())
-}
-
-#[tauri::command]
-pub fn browser_navigate(
-    app: AppHandle,
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    label: Option<String>,
-    url: String,
-    x: f64,
-    y: f64,
-    w: f64,
-    h: f64,
-    read_only_url: Option<String>,
-    sequence: u64,
-) -> Result<(), String> {
-    ensure_browser_controller(&caller)?;
-    let label = safe_browser_label(label);
-    Url::parse(&url).map_err(|e| e.to_string())?;
-    if let Some(read_only_url) = read_only_url.as_deref() {
-        Url::parse(read_only_url).map_err(|e| e.to_string())?;
-    }
-    if !x.is_finite() || !y.is_finite() || !w.is_finite() || !h.is_finite() {
-        return Err("browser bounds must be finite".to_string());
-    }
-    let lifecycle = lifecycle.inner().clone();
-    let bounds = BrowserBoundsIntent {
-        sequence,
-        x,
-        y,
-        w,
-        h,
-    };
-    {
-        let mut inner = lifecycle.lock()?;
-        if !record_navigation_intent(&mut inner, &label, sequence, url, read_only_url, bounds) {
-            return Ok(());
-        }
-    }
-    schedule_browser_reconcile(app, lifecycle, label);
-    Ok(())
-}
-
-#[tauri::command]
-pub fn browser_set_bounds(
-    app: AppHandle,
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    label: Option<String>,
-    x: f64,
-    y: f64,
-    w: f64,
-    h: f64,
-    sequence: u64,
-) -> Result<(), String> {
-    ensure_browser_controller(&caller)?;
-    if !x.is_finite() || !y.is_finite() || !w.is_finite() || !h.is_finite() {
-        return Err("browser bounds must be finite".to_string());
-    }
-    let label = safe_browser_label(label);
-    let lifecycle = lifecycle.inner().clone();
-    let bounds = BrowserBoundsIntent {
-        sequence,
-        x,
-        y,
-        w,
-        h,
-    };
-    {
-        let mut inner = lifecycle.lock()?;
-        if !record_bounds_intent(&mut inner, &label, bounds) {
-            return Ok(());
-        }
-    }
-    schedule_browser_reconcile(app, lifecycle, label);
-    Ok(())
-}
-
-#[tauri::command]
-pub fn browser_hide(
-    app: AppHandle,
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    label: Option<String>,
-    sequence: u64,
-) -> Result<(), String> {
-    ensure_browser_controller(&caller)?;
-    let label = safe_browser_label(label);
-    let lifecycle = lifecycle.inner().clone();
-    {
-        let mut inner = lifecycle.lock()?;
-        if !record_visibility_intent(&mut inner, &label, sequence, BrowserVisibility::Hidden) {
-            return Ok(());
-        }
-    }
-    schedule_browser_reconcile(app, lifecycle, label);
-    Ok(())
-}
-
-#[tauri::command]
-pub fn browser_hide_all_except(
-    app: AppHandle,
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    label: Option<String>,
-    sequence: u64,
-) -> Result<(), String> {
-    ensure_browser_controller(&caller)?;
-    let keep = label.map(|raw| safe_browser_label(Some(raw)));
-    schedule_scope_reconcile(
-        app,
-        lifecycle.inner().clone(),
-        BROWSER_LABEL_PREFIX.to_string(),
-        sequence,
-        BrowserScopeAction::Hide,
-        keep,
-    )
-}
-
-#[tauri::command]
-pub fn browser_close(
-    app: AppHandle,
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    label: Option<String>,
-    sequence: u64,
-) -> Result<(), String> {
-    ensure_browser_controller(&caller)?;
-    let label = safe_browser_label(label);
-    let lifecycle = lifecycle.inner().clone();
-    {
-        let mut inner = lifecycle.lock()?;
-        if !record_visibility_intent(&mut inner, &label, sequence, BrowserVisibility::Closed) {
-            return Ok(());
-        }
-    }
-    schedule_browser_reconcile(app, lifecycle, label);
-    Ok(())
-}
-
-fn pane_prefix(label: Option<String>) -> String {
-    match label {
-        Some(raw) => format!("{}-tab-", safe_browser_label(Some(raw))),
-        None => BROWSER_LABEL_PREFIX.to_string(),
-    }
-}
-
-/// Hide every native browser WebView belonging to a pane without destroying
-/// it. Surface changes use this command so WebView2 cannot capture clicks over
-/// another surface, while a rapid return can safely show the same live child
-/// instead of racing Tauri's asynchronous close/removal from the registry.
-#[tauri::command]
-pub fn browser_deactivate_all(
-    app: AppHandle,
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    label: Option<String>,
-    sequence: u64,
-) -> Result<(), String> {
-    ensure_browser_controller(&caller)?;
-    let prefix = pane_prefix(label);
-    schedule_scope_reconcile(
-        app,
-        lifecycle.inner().clone(),
-        prefix,
-        sequence,
-        BrowserScopeAction::Hide,
-        None,
-    )
-}
-
-/// Destroy every native browser WebView belonging to a pane (labels look like
-/// `cave-browser-<pane>-tab-<id>`), or every cave-browser WebView when no pane
-/// label is given. Ordinary surface changes use browser_deactivate_all; this
-/// command is reserved for lifecycle points that truly require destruction.
-#[tauri::command]
-pub fn browser_close_all(
-    app: AppHandle,
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    label: Option<String>,
-    sequence: u64,
-) -> Result<(), String> {
-    ensure_browser_controller(&caller)?;
-    let prefix = pane_prefix(label);
-    schedule_scope_reconcile(
-        app,
-        lifecycle.inner().clone(),
-        prefix,
-        sequence,
-        BrowserScopeAction::Close,
-        None,
-    )
-}
-
-#[tauri::command]
-pub fn browser_reload(
-    app: AppHandle,
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    label: Option<String>,
-    sequence: u64,
-) -> Result<(), String> {
-    ensure_browser_controller(&caller)?;
-    let label = safe_browser_label(label);
-    let lifecycle = lifecycle.inner().clone();
-    {
-        let mut inner = lifecycle.lock()?;
-        if !record_reload_intent(&mut inner, &label, sequence) {
-            return Ok(());
-        }
-    }
-    schedule_browser_reconcile(app, lifecycle, label);
-    Ok(())
-}
-
-/// Marks the next child-initiated navigation with a generation newer than the
-/// page currently displayed. This is only an attribution hint; the command
-/// grants no navigation or lifecycle authority to the untrusted child page.
-#[tauri::command]
-pub fn browser_report_user_navigation(
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    target_url: String,
-    allow_query_change: bool,
-) -> Result<u64, String> {
-    let label = caller.label().to_string();
-    if !label.starts_with(BROWSER_LABEL_PREFIX) {
-        return Err("browser navigation reports require a browser child webview".to_string());
-    }
-    if target_url.len() > 4096 {
-        return Err("browser navigation target is too long".to_string());
-    }
-    let target = Url::parse(&target_url).map_err(|_| "invalid browser navigation target")?;
-    if !matches!(target.scheme(), "http" | "https") {
-        return Err("browser navigation target must use http or https".to_string());
-    }
-    let tracker = event_tracker_for_label(lifecycle.inner(), &label)?;
-    let mut tracker = tracker
-        .lock()
-        .map_err(|_| "browser event tracker lock poisoned".to_string())?;
-    Ok(tracker.begin_user_navigation(&target, allow_query_change))
-}
-
-/// Called by the injected script inside a child browser webview so the real
-/// document.title can be emitted as a `browser:title` event on the main
-/// app event bus (where the BrowserPane JS component can receive it).
-/// This avoids the cross-webview event delivery problem in Tauri v2.
-#[tauri::command]
-pub fn browser_report_title(
-    app: AppHandle,
-    lifecycle: State<'_, BrowserLifecycleState>,
-    caller: tauri::Webview,
-    title: String,
-) -> Result<(), String> {
-    let label = caller.label().to_string();
-    if !label.starts_with(BROWSER_LABEL_PREFIX) {
-        return Err("browser title reports require a browser child webview".to_string());
-    }
-    let url = caller.url().map_err(|error| error.to_string())?;
-    let sequence = event_sequence_for_label_url(lifecycle.inner(), &label, &url);
-    let url = url.to_string();
-    let title = title.chars().take(512).collect::<String>();
-    let _ = app.emit(
-        "browser:title",
-        BrowserTitleEvent {
-            label,
-            title,
-            url,
-            sequence,
-        },
-    );
-    Ok(())
-}
-
-#[tauri::command]
-pub fn browser_report_scroll(
-    app: AppHandle,
-    caller: tauri::Webview,
-    scroll_y: f64,
-) -> Result<(), String> {
-    let label = caller.label().to_string();
-    if !label.starts_with(BROWSER_LABEL_PREFIX) {
-        return Err("browser scroll reports require a browser child webview".to_string());
-    }
-    if !scroll_y.is_finite() {
-        return Err("browser scroll position must be finite".to_string());
-    }
-    let _ = app.emit(
-        "browser:scroll",
-        BrowserScrollEvent {
-            label,
-            scroll_y: scroll_y.clamp(0.0, 1_000_000_000.0),
-        },
-    );
     Ok(())
 }
